Create Standard style on startup (gets overriden by templates and docs) Don't crash in KWTextParag::loadLayout if the style can't be found (shaheed!) Fixed jpmartin bug #1, copyParagData. (oh boy, there are 20!)
svn path=/trunk/koffice/; revision=90341

+++ kword/kwimage.cc
@@ -43,9 +43,9 @@ void KWTextImage::adjustToPainter( QPainter* )
     if ( !m_image.isNull() ) {
         KWDocument * doc = static_cast<KWTextDocument *>(parent)->textFrameSet()->kWordDocument();
         width = m_image.originalSize().width();
-        width = doc->zoomItX( (double)width ) / POINT_TO_INCH( QPaintDevice::x11AppDpiX() );
+        width = (int)( doc->zoomItX( (double)width ) / POINT_TO_INCH( QPaintDevice::x11AppDpiX() ) );
         height = m_image.originalSize().height();
-        height = doc->zoomItY( (double)height ) / POINT_TO_INCH( QPaintDevice::x11AppDpiY() );
+        height = (int)( doc->zoomItY( (double)height ) / POINT_TO_INCH( QPaintDevice::x11AppDpiY() ) );
         // This ensures 1-1 at 100% on screen, but allows zooming and printing with correct DPI values
         kdDebug() << "KWTextImage::adjustToPainter scaling to " << width << ", " << height << endl;
         m_image = m_image.scale( QSize( width, height ) );
